% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actual_values.R
\name{actual_values}
\alias{actual_values}
\title{Check whether the formulae for mu and sigma give anticipated c and p}
\usage{
actual_values(p, c, fc = 1)
}
\arguments{
\item{p}{The anticipated prevalence}

\item{c}{The anticipated C-statistic}

\item{fc}{The tuning factor to adjust mu and sigma when C>=0.8}
}
\value{
the actual prevalence and C-statistic, mu and sigma
}
\description{
The purpose of this function is to Check whether the values of mu and sigma
correspond to the anticipated values of prevalence, p, and C-statistic, c
that are arguments for the sample size calculation. These are relevant for
the version of the formulae that require the use of numerical integration
}
