% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size_c_app.R
\name{size_c_app}
\alias{size_c_app}
\title{Precision-based sample size calculation based on the C-statistic (C) -
approximation}
\usage{
size_c_app(p, c, se_c)
}
\arguments{
\item{p}{(numeric) The outcome prevalence, a real number between 0 and 0.5}

\item{c}{(numeric) The C-statistic,  a real number between 0.5 and 1}

\item{se_c}{(numeric) The required standard error of the estimated
C- statistic}
}
\value{
n The required sample size
}
\description{
The purpose of this function is to receive the anticipated outcome prevalence
, C-statistic and calculate the sample size to achieve the required precision
based on the C-statistic.
}
\examples{
size_c_app(0.057, 0.77, 0.025)

}
