% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size_cil.R
\name{size_cil}
\alias{size_cil}
\title{Precision-based sample size calculation based on the Calibration in the Large
(CiL)}
\usage{
size_cil(p, c, varcl)
}
\arguments{
\item{p}{(numeric) The outcome prevalence, a real number between 0 and 0.5}

\item{c}{(numeric) The C-statistic, a real number between 0.5 and 1}

\item{varcl}{(numeric) The required variance of the Calibration in the Large}
}
\value{
n The required sample size
}
\description{
Precision-based sample size calculation based on the Calibration in the Large
(CiL)
}
\examples{
size_cil(0.057, 0.77, 0.15^2)

}
