% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size_cil_app.R
\name{size_cil_app}
\alias{size_cil_app}
\title{Precision-based sample size calculation based on the Calibration in the Large
(CiL) - approximation}
\usage{
size_cil_app(p, c, se_cl)
}
\arguments{
\item{p}{(numeric) The outcome prevalence, a real number between 0 and 0.5}

\item{c}{(numeric) The C-statistic, a real number between 0.5 and 1}

\item{se_cl}{(numeric) The required standard error of the estiamted
Calibration in the Large}
}
\value{
n The required sample size
}
\description{
Precision-based sample size calculation based on the Calibration in the Large
(CiL) - approximation
}
