% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size_cs.R
\name{size_cs}
\alias{size_cs}
\title{Precision-based sample size calculation based on the Calibration Slope (CS)}
\usage{
size_cs(p, c, varcs)
}
\arguments{
\item{p}{(numeric) The outcome prevalence, a real number between 0 and 0.5}

\item{c}{(numeric) The C-statistic, a real number between 0.5 and 1}

\item{varcs}{(numeric) The required variance of the Calibration Slope}
}
\value{
n The required sample size
}
\description{
The purpose of this function is to receive the anticipated outcome prevalence
, C-statistic and calculate the sample size to achieve the required precision
based on CS.
}
\examples{
size_cs(0.057, 0.77, 0.15^2)

}
