% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size_cs_app.R
\name{size_cs_app}
\alias{size_cs_app}
\title{Precision-based sample size calculation based on the Calibration Slope (CS) -
approximation}
\usage{
size_cs_app(p, c, se_cs)
}
\arguments{
\item{p}{(numeric) The outcome prevalence, a real number between 0 and 0.5}

\item{c}{(numeric) The C-statistic, a real number between 0.5 and 1}

\item{se_cs}{(numeric) The required standard error of the estimated
Calibration Slope}
}
\value{
n The required sample size
}
\description{
The purpose of this function is to receive the anticipated outcome prevalence
, C-statistic and calculate the sample size to achieve the required precision
based on CS.
}
\examples{
size_cs(0.057, 0.77, 0.15)

}
