% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size_val.R
\name{size_val}
\alias{size_val}
\title{Precision-based sample size calculations for the validation of risk models
for binary outcomes}
\usage{
size_val(p, c, se_c = 0.025, se_cs = 0.1, se_cl = 0.1)
}
\arguments{
\item{p}{(numeric) The outcome prevalence, a real number between 0 and 0.5}

\item{c}{(numeric) The C-statistic, a real number between 0.5 and 1}

\item{se_c}{(numeric) The required standard error of the estimated
C-Statistic}

\item{se_cs}{(numeric) The required standard error of the estimated
Calibration Slope}

\item{se_cl}{(numeric) The required standard error of the estimated
Calibration in the large}
}
\value{
size_c: the sample size based on the C-statistic,

size_cs: the sample size based on the Calibration Slope,

size_cl: the sample size based on the Calibration in the Large and the final recommended sample size

size_recommended: the final sample size recommendation
}
\description{
This function calculates the sample size required in the validation dataset
to estimate the C-statistic (C), the calibration Slope (CS) and the
Calibration in the Large (CL) with sufficient precision. It takes as
arguments the anticipated values of the C-statistic and the outcome
prevalence(obtained, for example, from a previous study) and the required
standard error for C, CS and CL.
}
\details{
The sample size calculations are valid under the assumption of marginal
normality for the distribution of the linear predictor.The default sample
size calculation based on C uses the closed-form expression in equation (9).
This is quick to run and accurate for all values of C and p.
The default sample size calculation based on CS and CL
use the formulae (12) and (13) that require the use of numerical integration.
These are also quick to run.

Results from the estimator based on C that uses numerical integration #' and
the closed-form estimators based on CS and CL that make the use of
approximations can be obtained from functions  'size_c_ni', 'size_cs_app'
and 'size_cl_app'.
}
\examples{
size_val(p=0.1, c=0.75, se_c=0.025, se_cs =0.1, se_cl = 0.1)
}
