% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_mu_sigma.R
\name{tune_mu_sigma}
\alias{tune_mu_sigma}
\title{Find mu and sigma that correspond to antcipated C and p (when C>=0.8)}
\usage{
tune_mu_sigma(p, c)
}
\arguments{
\item{p}{The anticipated outcome prevalence}

\item{c}{The anticipated C-statistic}
}
\value{
The actual prevalence and c-statistic, for the updated values of
mu and sigma based on the tuning factor fc.
}
\description{
The purpose of this function is to check whether the values of mu
and sigma as given by the formulae in the paper
give actual value of C sufficiently close to the anticipated
values. If not, mu and sigma are fine-tuned to provide actual values of c and
p sufficiently close to the anticipated ones. This is relevant for the
version of the formulae that require the use of numerical integration, when
C>=0.8.
}
