% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size_c_ni.R
\name{size_c_ni}
\alias{size_c_ni}
\title{Precision-based sample size calculation based on the C-statistic (C)
using Numerical Integration}
\usage{
size_c_ni(p, c, varc)
}
\arguments{
\item{p}{(numeric) The outcome prevalence, a real number between 0 and 0.5}

\item{c}{(numeric) The C-statistic, a real number between 0.5 and 1}

\item{varc}{(numeric) The required variance of the C-Statistic}
}
\value{
n         The required sample size
}
\description{
The purpose of this function is to receive the anticipated outcome prevalence
, C-statistic and calculate the sample size to achieve the required precision
based on the C-Statistic. It uses Numerical integration and it assumes
marginal normality.
}
\examples{
size_c_ni(0.057, 0.77, 0.025^2)

}
