% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size_cs_ni.R
\name{size_cs_ni}
\alias{size_cs_ni}
\title{Precision-based sample size calculation based on the Calibration Slope (CS) -
Numerical Integration}
\usage{
size_cs_ni(p, c, se_cs)
}
\arguments{
\item{p}{(numeric) The outcome prevalence, a real number between 0 and 0.5}

\item{c}{(numeric) The C-statistic, a real number between 0.5 and 1}

\item{se_cs}{(numeric) The required variance of the Calibration slope}
}
\value{
n The required sample size
}
\description{
The purpose of this function is to receive the anticipated outcome prevalence
, -statistic and calculate the sample size to achieve the required precision
based on CS. It uses Numerical integration and it assumes marginal normality.
}
