% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size_val.R
\name{size_val}
\alias{size_val}
\title{Precision-based sample size calculations for the validation of risk models
for binary outcomes}
\usage{
size_val(p, c, se_c = 0.025, se_cs = 0.1, se_cl = 0.1)
}
\arguments{
\item{p}{(numeric) The outcome prevalence, a real number between 0 and 0.5}

\item{c}{(numeric) The C-statistic, a real number between 0.5 and 1}

\item{se_c}{(numeric) The required standard error of the estimated
C-Statistic}

\item{se_cs}{(numeric) The required standard error of the estimated
Calibration Slope}

\item{se_cl}{(numeric) The required standard error of the estimated
Calibration in the large}
}
\value{
size_c: the sample size based on the C-statistic,

size_cs: the sample size based on the Calibration Slope,

size_cl: the sample size based on the Calibration in the Large and the final recommended sample size

size_recommended: the final sample size recommendation
}
\description{
This function calculates the sample size required in the validation dataset
to estimate the C-statistic (C), the calibration Slope (CS) and the
Calibration in the Large (CL) with sufficient precision. It takes as
arguments the anticipated values of the C-statistic and the outcome
prevalence(obtained, for example, from a previous study) and the required
standard error for C, CS and CL. The calculations are valid under the
assumption of marginal
normality for the distribution of the linear predictor.
}
\examples{
size_val(p=0.1, c=0.75, se_c=0.025, se_cs =0.1, se_cl = 0.1)
}
